发现还有地方没改

+++ src/main/java/com/zjgsu/shopping/service/SellerService.java
@@ -116,7 +116,7 @@ public interface SellerService {
     /**
      * 上架一个商品
      */
-    GoodForSale putOnGood(String name,String description,Integer price);
+    GoodForSale putOnGood(String name,String description,Double price);
 
 
     GoodForSale putOnGood(GoodForSale good);
